+++ src/main/java/nl/dtls/fairdatapoint/api/dto/settings/SettingsUpdateDTO.java
@@ -35,6 +35,10 @@ import java.util.List;
 @Builder(toBuilder = true)
 public class SettingsUpdateDTO {
 
+    private String appTitle;
+
+    private String appSubtitle;
+
     @NotNull
     private List<SettingsMetricsEntry> metadataMetrics;